Revert "Merge branch 'feature/JNG-2630_Col_value_calculation'"
This reverts commit 6d0cd6a101872ebecee1f8af8a5b012e82a992c1, reversing
changes made to 792ec8fd76bdb804be9cdcc86c914fcc6a487eb9.

+++ lib/src/judo_link.dart
@@ -18,7 +18,7 @@ class JudoLink extends StatelessWidget {
     this.alignment = Alignment.centerLeft,
   }) : super(key: key);
 
-  final double col;
+  final int col;
   final double row;
   final bool disabled;
   final bool readOnly;
