+++ lib/src/judo_numeric_input.dart
@@ -19,7 +19,7 @@ class JudoNumericInput extends StatefulWidget {
     this.alignment = Alignment.centerLeft,
   }) : super(key: key);
 
-  final double col;
+  final int col;
   final double row;
   final bool mandatory;
   final String label;
